fix: progressbar without animation from beginning and questions counter
